fix cypress issue
