+++ web/cypress/integration/Home/CancelTest.spec.ts
@@ -1,0 +1,8 @@
+describe('Cancel Create test', () => {
+  beforeEach(() => cy.visit('/'));
+
+  it('should cancel a create test flow', () => {
+    cy.navigateToTestCreationPage();
+    cy.get('[data-cy=create-test-cancel]').click();
+  });
+});
